Created states module
- Refactored and stored class definitions in `definitions.py`
